made rc6 need to make some changes

+++ src/main/java/com/company/cripto/aesImpl/round/RoundKeysGenerator.java
@@ -7,5 +7,5 @@ import java.math.BigInteger;
  * Generate all round keys
  **/
 public interface RoundKeysGenerator {
-    long[] generate(byte[] cipherKey);
+    int[] generate(byte[] cipherKey);
 }
